Fix broken discobeat section in Crowded Sky
Something is really messed up with "drums0d" events.
I already knew that the note simultaneous with the text event
is not affected by discobeat.
But apparently, even if you *already are in discobeat* and you have
a redundant "drums0d" event (which I did for easier Reaper editing),
the note simultaneous with the event is *still is not in discobeat*.
Harmonix how did you screw this up so bad?