Fix yet another windows utf-8 crash
Short path hack only works if short names are enabled on the drive.
On a non C: drive they likely aren't, so importing from a non-Latin-1 character causes ffmpeg to crash.
There's no consistent C-side fix (UTF-8 fopen exists in later Windows 10 versions but can't rely on it) so we open the file on the Haskell side instead.