Maybe fix a further temp deletion issue on Windows
When importing from a RB file with a mogg, the current process
first decodes the ogg into a wav, and then any new files that
require it (like separate CH oggs) will all read channels at once
from the wav. This appears to cause some problems on Windows where
the wav fails to be deleted when there were several read handles
attached to it? Or something like that

So to alleviate, we decode the ogg into one-channel wavs up front.
Need to test if this actually avoids the problem